requirements-dev.txt
black==21.9b0
click==8.0.2
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pyaes==1.6.1
pyasn1==0.4.8
python-decouple==3.5
regex==2021.10.8
rsa==4.7.2
Telethon==1.23.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
